moved Info component to new file
