+++ client/components/ArtistPage.js
@@ -7,7 +7,8 @@ export default class ArtistPage extends React.Component {
 		super(props)
 
 		this.state = {
-			art: []
+			art: [],
+      showInfo: false,
 		}
 	}
 
@@ -34,13 +35,17 @@ export default class ArtistPage extends React.Component {
     return (
       <div>
         <NavBar />
+        {this.state.showInfo ?
+          <Info onClose={this.closeInfo.bind(this)} loggedIn={this.props.loggedIn} updateCurrent={this.updateCurrent.bind(this)} currentArt={this.state.currentArt} parseImageUrl={this.parseImageUrl.bind(this)}/>
+          : null}
         <br/>
         <br/>
         {this.state.art.map(art => {
           return (
             <div key={art._id} className="soloWork">
               <h3 className="soloArtTitle">{art['Art Title']}</h3>
-              <img className='artImage' src={this.parseImageUrl(art.Images)[0]} />
+              <a href="javascript:void(0)" onClick={(e) => this.openInfo(art)} className="artImage">
+              <img className='artImage' src={this.parseImageUrl(art.Images)[0]} /></a>
               <div className="soloArtInfo">
                 <p>{art['Art Location Name']}</p>
               </div>
