+++ client/components/NavBar.js
@@ -213,5 +213,8 @@ class SignUpModal extends React.Component {
 }
 
 class FBModal extends React.Component{
-  
+
 }
+
+
+
